+++ matrix-processing/libs/matrix/include/matrix/kernels/random.cuh
@@ -1,4 +1,4 @@
 #pragma once
 
-void launchRandom(unsigned char* devMemory, int rows, int columns);
+void launchRandom(unsigned char* devMemory, long long rows, long long columns);
 __global__ void random(unsigned char* devMemory);
