+++ matrix-processing/libs/matrix/include/matrix/kernels/shared.cuh
@@ -1,3 +1,3 @@
 #pragma once
 
-__global__ void reserve(unsigned char* devMemory, int n);
+__global__ void reserve(unsigned char* devMemory, long long n);
