+++ matrix-processing/libs/matrix/src/matrix-cpu.cpp
@@ -3,21 +3,21 @@
 #include <cmath>
 #include <random>
 
-MatrixCPU::MatrixCPU(int n, int m) : Matrix(n, m) {
+MatrixCPU::MatrixCPU(long long n, long long m) : Matrix(n, m) {
 }
 
 void MatrixCPU::process() {
-    int n = rows*2;
-    int m = columns/2;
+    long long n = rows*2;
+    long long m = columns/2;
     auto* result = new unsigned char[rows * columns];
 
-    for (int i = 0; i < rows; i++) {
+    for (long long i = 0; i < rows; i++) {
         auto rowsOffset = i*columns;
         auto doubleRowsOffset = rowsOffset;
-        for (int j = 0; j < columns; j += 4) {
+        for (long long j = 0; j < columns; j += 4) {
             if (j + 4 > columns)
                 continue;
-            auto startResultOffset = doubleRowsOffset + (int)std::ceil(j/2);
+            auto startResultOffset = doubleRowsOffset + (long long)std::ceil(j/2);
             result[startResultOffset] = matrix[rowsOffset + j + 2];
             result[startResultOffset+1] = matrix[rowsOffset + j + 3];
             result[startResultOffset+m] = matrix[rowsOffset + j];
@@ -35,8 +35,8 @@ void MatrixCPU::initialize() {
     std::default_random_engine generator;
     std::uniform_int_distribution<unsigned char> distribution(10,100);
 
-    for (int i = 0; i < rows; i++) {
-        for (int j = 0; j < columns; j++) {
+    for (long long i = 0; i < rows; i++) {
+        for (long long j = 0; j < columns; j++) {
             matrix[i*columns+j] = distribution(generator);
         }
     }
